feat: 사용자 타입 정의를 '@/shared/types/user'로 이동 및 관련 코드 수정
