--- src/features/auth/api-client.ts
@@ -1,5 +1,0 @@
-/**
- * Auth feature API client
- * Authentication 관련 API들은 대부분 public이므로 fetchApi 사용
- */
-export { fetchApi as api } from '@/shared/api/fetch-api'
